moved collision files into collision directory (i want to keep it all together) additionally fixed some instances of absolute time being stored in a float
